stage: fix stage_build_output scheme
`cp -u` copy only when the source file is newer than the destination.
This is not the assumption that should be made for the staging step,
stage_iso uses symlinks as sources in some cases (e.g, kernel, initrd).

If a symlink it moved from one file to the other due to version changes
(notably kernel binaries), it can be reverted to a timestamp of an old
kernel that is still present in the DEPLOY_DIR, but will not be copied
to the staging again.

The staging should compare files, since it is also supposed to maintain
symlinks and directory hierarchies, the easiest is to rely on rsync to
do so. This also simplifies suffix matching.

While modifying stage_build_output, also amend the sanity checks.

Signed-off-by: Eric Chanudet <chanudete@ainfosec.com>